Move controller_test.go into controller.go
Turns out the stuff in the test file needs to be exported so external add-ons can use them
